v2.1.0 - closes #9, closes #11
