+++ .gitignore
@@ -3,3 +3,7 @@
 .idea
 logs
 .env
+*.egg-info
+build
+dist
+__pycache__
