--- requirements.txt
@@ -1,3 +1,0 @@
-python-dotenv
-PyGithub
-pylint
