--- src/.env.example
@@ -1,44 +1,0 @@
-## GITHUB ARCHIVE CONFIGURATION ##
-# Only edit the values of each variable, leaving variable names alone.
-
-######################################################################
-## USER CONFIGURATION ##
-
-# Your GitHub token
-TOKEN=""
-
-# Enter "enable" to clone personal repos. Default: "enable"
-USER_CLONE_ON="enable"
-
-# Enter "enable" to pull personal repos. Default: "enable"
-USER_PULL_ON="enable"
-
-# Enter "enable" to clone gists. Default "disable"
-GISTS_CLONE_ON="disable"
-
-# Enter "enable" to pull gists. Default "disable"
-GISTS_PULL_ON="disable"
-
-######################################################################
-## ORGANIZATIONAL CONFIGURATION ##
-
-# Add each org name separated by comma. eg: "org1, org2, org3"
-ORGS=""
-
-# Enter "enable" to clone organization repos. Organization names must be entered in the "ORGS" variable below. Default: "disable"
-ORGS_CLONE_ON="disable"
-
-# Enter "enable" to pull organization repos. Organization names must be entered in the "ORGS" variable below. Default: "disable"
-ORGS_PULL_ON="disable"
-
-######################################################################
-## ARCHIVE CONFIGURATION ##
-
-# Number of days logs will be retained for. Default: "30"
-LOG_LIFE="30"
-
-# Where your archive will be housed. Default: "~/github-archive"
-LOCATION="~/github-archive"
-
-# Which branch each repo will pull from. Default: "master"
-BRANCH="master"
