--- src/local.githubArchive.plist
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-    <dict>
-		<key>Label</key>
-		<string>local.githubArchive.plist</string>
-		<key>ProgramArguments</key>
-		<array>
-			<string>/usr/local/bin/python3</string>
-			<string>/path/to/github_archive.py</string>
-		</array>
-		<key>StartCalendarInterval</key>
-		<dict>
-			<key>Hour</key>
-			<integer>1</integer>
-			<key>Minute</key>
-			<integer>0</integer>
-		</dict>
-		<key>StandardErrorPath</key>
-		<string>/path/to/github-archive/logs/log.err</string>
-		<key>StandardOutPath</key>
-		<string>/path/to/github-archive/logs/log.out</string>
-	</dict>
-</plist>
